Added support for out-of-source class path.
Now the adcv.cls file (alongside the fonts and icons directories)
can be placed on a separate path than the main document directory.

For example, another person could write a custom CV file and clone the
adcv project in a subdirectory (or use a git submodule),
then use the following command to set the document class:
`\documentclass[...]{relpath/adcv}`

Before this change, all paths referred to in adcv.cls broke if someone attempted to do
this.

In addition, it is now possible to place adcv_lang.def files in the main
document directory, which may be different than the adcv.cls
file directory thanks to the changes in this commit.